requirements.txt
fastapi==0.109.0
googlemaps==4.10.0
pydantic==2.5.3
pydantic_core==2.14.6
uvicorn==0.26.0
Werkzeug==3.0.1
zipp==3.17.0
Jinja2==3.1.3
json2table==1.1.5
requests==2.31.0
urllib3==2.2.0
mysql-connector==2.2.9